requirements.txt
click==8.0.3
Flask==2.0.2
flask-sqlalchemy==2.5.1
SQLAlchemy
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
PyMySQL==1.0.2
SQLAlchemy==1.4.29
Werkzeug==2.0.2
pyjwt == 2.4.0
bcrypt == 4.0.1
Pillow~=9.5.0
cos-python-sdk-v5
gunicorn == 21.2.0